requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
apimoex==1.3.0
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
bleach==4.1.0
build==0.10.0
CacheControl==0.12.11
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cleo==2.0.1
colorama==0.4.4
coloredlogs==15.0.1
crashtest==0.4.1
cryptography==35.0.0
distlib==0.3.6
dnspython==2.3.0
docutils==0.18
dulwich==0.21.5
filelock==3.12.0
frozenlist==1.3.3
html5lib==1.1
humanfriendly==10.0
idna==3.3
importlib-metadata==4.8.1
importlib-resources==5.12.0
installer==0.7.0
jaraco.classes==3.2.3
jeepney==0.7.1
jsonschema==4.17.3
keyring==23.2.1
lockfile==0.12.2
loguru==0.5.3
more-itertools==9.1.0
motor==3.1.2
msgpack==1.0.5
mt5linux==0.1.9
multidict==6.0.4
nodeenv==1.8.0
numpy==1.21.4
orjson==3.8.14
packaging==21.2
pexpect==4.8.0
pkginfo==1.7.1
pkgutil_resolve_name==1.3.10
platformdirs==3.5.1
plumbum==1.7.0
ptyprocess==0.7.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.8
Pygments==2.10.0
pymongo==4.3.3
pyparsing==2.4.7
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dotenv==1.0.0
PyYAML==6.0
rapidfuzz==2.15.1
readme-renderer==30.0
requests==2.26.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rpyc==5.0.1
SecretStorage==3.3.1
shellingham==1.5.0.post1
six==1.16.0
tomli==2.0.1
tomlkit==0.11.8
tqdm==4.62.3
trove-classifiers==2023.5.24
twine==3.5.0
typing_extensions==4.6.2
urllib3==1.26.7
virtualenv==20.23.0
webencodings==0.5.1
yarl==1.9.2
zipp==3.6.0
clickhouse-connect==0.6.23
clickhouse-driver==0.2.7
psycopg2-binary==2.9.9
